removed initEventList(), changed interface to print event, print event list functions
git-svn-id: http://atariforge.org/svn/amidilib/trunk@186 5b2f37e0-8fbc-4d39-81b1-0c51156b4091

+++ main.c
@@ -71,7 +71,7 @@ int main(int argc, char *argv[]){
     pMidi=loadFile((U8 *)argv[1], PREFER_TT, &ulFileLenght);
 
     if(pMidi!=NULL){
-      fprintf(stderr,"Midi file loaded, size: %u bytes.\n",(unsigned int)ulFileLenght);
+      fprintf(stderr,"Midi file loaded, size: %u bytes.\n",(unsigned long)ulFileLenght);
      
      /* process MIDI*/
      /* check midi header */
